Exposed C++ Class Config

Analog generation can be fine-tuned with configuration. This class provides all the tunable parameters for analog generation. You can type RAnEn::Config in the R console to see the description for each parameter. To create an object of Config, you can type config <- new(Config), and then you can directly modify the members, e.g. config$num_analogs <- 10. To see the full list of tunable parameters, you can type config after you have created an object. It will print all the parameters.