RAnEn::RAnEn

The Analog Ensemble Package computes analog ensembles. It is implemented in C++ and exported to R interface.

Details

RAnEn is the R package implementing the Analog Ensemble method and more with data processing and visualization tasks.

See also

Author

Weiming Hu weiming@psu.edu

Guido Cervone cervone@psu.edu

Martina Calovi mxc895@psu.edu

Laura Clemente-Harding laura@psu.edu