RAnEn::generateHPCWorkflow

RAnEn::generateHPCWorkflow can generate all configuration files for an HPC workflow. If your Analog Ensemble generation requires supercomputers, this is a very helpful function to generate the configuration files for each computational node.

generateHPCWorkflow(
  test.start,
  test.end,
  search.start,
  search.end,
  fcst.files,
  obs.files,
  num.members,
  output.folder,
  batch.stations,
  batch.times,
  obs.id = 1,
  weights = NULL,
  verbose = 3,
  max.par.nan = 1,
  max.flt.nan = 1,
  max.num.sims = 2 * num.members,
  overwrite = F
)

Arguments

test.start
A POSIXct for the start test time

test.end
A POSIXct for the end test time

search.start
A POSIXct for the start search time

search.end
A POSIXct for the end search time

fcst.files
The forecast file(s)

obs.files
The observation file(s)

num.members
The number of members in each analog

output.folder
The output folder path

batch.stations
How many stations to process in each node

batch.times
The number of seconds included in each node. This is used to parallel along time. For example, you have 3 test days and you want to process each day separately, you can set batch.times to 24 * 60 * 60.

obs.id
The observation id

weights
The weights for each forecast variable

verbose
Verbose level. Please see this doc

max.par.nan
Maximum number of NAN in parameters.

max.flt.nan
Maximum number of NAN in FLTs.

max.num.sims
How many similarity metrics to keep in the file.

overwrite
Whether to overwrite existing configuration files.

Details

If forecast and observations are stored in multiple files, it is assumed that they are splitted by time.

Examples

if (FALSE) {
test.start <- as.POSIXct('2018-08-09', tz = 'UTC')
test.end <- as.POSIXct('2018-08-19', tz = 'UTC')
search.start <- as.POSIXct('2017-01-20', tz = 'UTC')
search.end <- as.POSIXct('2017-12-15', tz = 'UTC')

fcst.files <- list.files(
  path = '/glade/u/home/wuh20/scratch/data/AnEn/forecasts',
  pattern = paste('2017|2018', collapse = ''),
  full.names = T)
obs.files <- list.files(
  path = '/glade/u/home/wuh20/scratch/data/AnEn/analysis',
  pattern = paste('2017|2018', collapse = ''),
  full.names = T)

max.flt <- 302400
output.folder <- '.'
num.members <- 20

batch.stations <- 1000
batch.times <- 24 * 3600 * 4

generateHPCWorkflow(
  test.start = test.start,
  test.end = test.end,
  search.start = search.start,
  search.end = search.end,
  fcst.files = fcst.files,
  obs.files = obs.files,
  num.members = num.members,
  output.folder = output.folder,
  batch.stations = batch.stations,
  batch.times = batch.times,
  overwrite = T)
}