RAnEn::generateTimeMapping

RAnEn::generateTimeMapping returns a matrix that maps each combination of forecast times and flts to observation times. The rows are forecast FLTs and columns are forecast times, the values are the index of the corresponding observation times. Same results can be generated from the function generateConfiguration if preserve_mapping is selected.

generateTimeMapping(forecast_times, forecast_flts, observation_times,
  time_match_mode = 0, verbose = 2)

Arguments

forecast_times
A vector of forecast times.

forecast_flts
A vector of forecast FLTs.

observation_times
A vector of observation times.

time_match_mode
This is the method used to computed the mapping matrix between forecast times/flts and observation times. 0 for strict search (return error when matching observation times cannot be found) and 1 for loose search.

verbose
Verbose level. You can find settings for different values in the [C++ documentation verbose](https://weiming-hu.github.io/AnalogsEnsemble/CXX/class_an_en.html#a25984b953516a987e2e9eb23048e5d60).

Value

A matrix with rows for forecast FLTs and columns for forecast times.