RAnEn::getAssociatedForecasts

RAnEn::getAssociatedForecasts returns the most similar past forecasts that are identified using this analog ensemble and associated with each analog ensemble member. In other words, this funtion returns the most similar past forecasts to the current forecasts.

getAssociatedForecasts(AnEn, config, forecast.parameter, i.station = NA,
  i.test.day = NA, i.flt = NA, similarity.name = "similarity",
  forecasts.name = "forecasts", members.to.extract = NA)

Arguments

AnEn
The AnEn class object returned by generateAnalogs.

config
The Configuration class object used to generate AnEn.

forecast.parameter
The parameter index for which variable to extract in the forecast array in config.

i.station
The station index in AnEn to extract.

i.test.day
The test day index in AnEn to extract.

i.flt
The FLT index in AnEn to extract.

similarity.name
The similarity name to be used in AnEn. This is useful when you have a different similarity matrice name to the regular one.

forecasts.name
The forecast name to be used in config. This is useful when you have a different similarity matrice name to the regular one.

members.to.extract
The number of ensemble members to extract. By default to extract all members.

Value

An array with the same dimension as analogs in AnEn. For more information of each dimension, please refer to generateAnalogs. The only difference is that the values in this array are extracted from the most similar past forecast that is associated with each analog ensemble member.