RAnEn::toForecastTime

RAnEn::toForecastTime converts the observation time index to forecast time index.

toForecastTime(observation.time.id, mapping, flt = NULL)

Arguments

observation.time.id

The observation time index(es) from the Time column of the analogs member in in the results of generateAnalogs.default.

mapping

The mapping table created from generateTimeMapping.

flt

The lead time index.

Details

This function does the following conversion.

Observation time index [ + Forecast lead time index ] ----> Forecast time index

To pin down the exact time of a forecast, two things are needed: forecast time and the lead time; to pin down the exact time of an observation, only the observation time is needed. Therefore, this function is devoted to convert from observation times to forecast times. Please note that, an observation time CAN be associated with multiple forecast times because of the existence of forecast lead times. If you have constraints on the lead time, then the match between forecasts and observations becomes one-on-one.

The observation time index usually comes from the Time column of the analogs member in the results of generateAnalogs.default. This time is observation times. However, please note that the times in the similarity member are forecast times. These two are different.

Author

Guido Cervone cervone@psu.edu

Martina Calovi mxc895@psu.edu

Laura Clemente-Harding laura@psu.edu