RAnEn::verifyMAE

RAnEn::verifyMAE calculates mean absolute errors.

verifyMAE(anen.ver, obs.ver, boot = F, R = 1000, na.rm = T,
  parallel = F)

Arguments

anen.ver
A 4-dimensional array. This array is usually created from the value column of the analogs member in the results of generateAnalogs. The dimensions should be [stations, times, lead times, members].

obs.ver
A 3-dimensional array. The dimensions should be [stations, times, lead times]. You can generate the array using alignObservations.

boot
Whether to use bootstrap.

R
The number of bootstrap replicates. Used by the function boot::boot.

na.rm
Whether to remove NA values.

parallel
Whether to use parallel processing.