RAnEn::RAnEn

The Analog Ensemble Package computes analog ensembles. It is implemented in C++ and exported to R interface.

Details

RAnEn is the R package implementing the Analog Ensemble method and more with data processing and visualization tasks.

See also